+++ Thunderstore/CHANGELOG.md
@@ -1,4 +1,9 @@
-# v1.0.0-pre9
+# v1.0.0-pre10
+
+- Further Improved DungeonGenerationPlus compatibility with Debug mode.
+- Added tooltips to MapTileSize and Dungeon Range Clamp on `DuskDungeonDefinition`.
+
+## v1.0.0-pre9
 
 - Added DawnInfo checks to prevent TileSets from re-registering. if found in multiple DungeonArchetypes.